projects/CMI_Detect_Behavior_with_Sensor_Data/codes/src/utils/seed_everything.py
import os
import random

import numpy as np
import pytorch_lightning as pl
import torch

try:
    import cupy as cp

    HAS_CUPY = True
except ImportError:
    HAS_CUPY = False


def seed_everything(seed=1234):
    random.seed(seed)
    os.environ["PYTHONHASHSEED"] = str(seed)
    np.random.seed(seed)
    if HAS_CUPY:
        cp.random.seed(seed)
    torch.manual_seed(seed)
    torch.cuda.manual_seed(seed)
    torch.backends.cudnn.deterministic = True
    pl.seed_everything(seed)
