projects/CMI_Detect_Behavior_with_Sensor_Data/requirements/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
alembic==1.16.4
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorlog==6.9.0
comm==0.2.3
contourpy==1.3.3
cupy-cuda12x==13.5.1
cycler==0.12.1
debugpy==1.8.15
decorator==5.2.1
defusedxml==0.7.1
einops==0.8.1
ema-pytorch==0.7.7
executing==2.2.0
fastjsonschema==2.21.1
fastrlock==0.8.3
filelock==3.18.0
fonttools==4.59.0
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.7.0
gitdb==4.0.12
gitpython==3.1.45
greenlet==3.2.3
grpcio==1.74.0
grpcio-tools==1.74.0
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.3
hydra-core==1.3.2
idna==3.10
imbalanced-learn==0.13.0
iniconfig==2.1.0
ipykernel==6.30.0
ipython==9.4.0
ipython-genutils==0.2.0
ipython-pygments-lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
jinja2==3.1.6
joblib==1.5.1
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter-client==8.6.3
jupyter-core==5.8.1
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter-server==2.16.0
jupyter-server-terminals==0.5.3
jupyterlab==4.3.8
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
kaggle==1.7.4.5
kagglehub==0.3.12
kiwisolver==1.4.8
lark==1.2.2
lightgbm==4.6.0
lightning==2.5.2
lightning-utilities==0.15.0
loguru==0.7.3
mako==1.3.10
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
matplotlib-venn==1.1.2
mdurl==0.1.2
mistune==3.1.3
mpmath==1.3.0
multidict==6.6.3
mypy==1.17.1
mypy-extensions==1.1.0
narwhals==2.0.1
nbclassic==1.3.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook==7.3.2
notebook-shim==0.2.4
numpy==2.3.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
opencv-python-headless==4.11.0.86
optuna==4.4.0
overrides==7.7.0
packaging==25.0
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
polars==1.32.0
prometheus-client==0.22.1
prompt-toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==21.0.0
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
pydantic-settings==2.10.1
pydeck==0.9.1
pygments==2.19.2
pyparsing==3.2.3
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-slugify==8.0.4
pytorch-lightning==2.5.2
pytz==2025.2
pyyaml==6.0.2
pyzmq==27.0.0
referencing==0.36.2
requests==2.32.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.1.0
rpds-py==0.26.0
ruff==0.12.7
safetensors==0.5.3
schedulefree==1.4.1
scikit-learn==1.6.1
scipy==1.16.1
seaborn==0.13.2
send2trash==1.8.3
sentry-sdk==2.34.1
setuptools==80.9.0
simsimd==6.5.0
six==1.17.0
sklearn-compat==0.1.3
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
sqlalchemy==2.0.42
stack-data==0.6.3
streamlit==1.47.1
stringzilla==3.12.5
sympy==1.13.1
tenacity==9.1.2
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.6.0
timm==1.0.19
tinycss2==1.4.0
toml==0.10.2
torch==2.6.0
torchmetrics==1.8.0
torchvision==0.21.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
triton==3.2.0
types-python-dateutil==2.9.0.20250708
typing-extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
wandb==0.21.0
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
yarl==1.20.1
